feat: add home page
To handle the authentication step i created a Home component that is 100% client side, so before returning that i check if there is auth token, for while im not checking if its valid.

We have a task list component that lists all tasks created by the user, inside of it we have a task component.

Also added a inputTask component, that allready register a task but need do be improved.