feat: add authorization to application
now the user is able to create an account and login, we got big changes into the AuthForm component, where i used react hook and zod to validate data.

I also used axios to make requests, and to save cookies i did a long work around, falling into the api of next to be able to save the cookies from server side.